fix: return QA
