--- components/DashboardHeader.css
@@ -1,91 +1,0 @@
-#dashboard-header .header {
-  background: url('/static/background.webp') center no-repeat;
-  background-size: cover;
-  height: 30vh;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-#dashboard-header .header .home-title {
-  margin: 1em auto;
-  padding: 1.5em;
-  background-color: var(--dark-color);
-  transform: skew(15deg);
-  filter: drop-shadow(0 0 0.75rem rgba(255, 255, 255, 0.2));
-}
-
-#dashboard-header .header .home-title p {
-  font-family: 'LemonMilk', sans-serif;
-  text-align: center;
-  color: white;
-  transform: skew(-15deg);
-  margin: 0;
-  font-size: 3rem;
-  letter-spacing: 1px;
-}
-
-#dashboard-header .nav {
-  background: var(--dark-color);
-  height: 50px;
-  display: flex;
-  flex-flow: row nowrap;
-  overflow-x: auto;
-}
-
-#dashboard-header .nav .nav-link {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-grow: 1;
-  text-decoration: none;
-  font-family: 'Montserrat', sans-serif;
-  color: #fff;
-  min-width: max-content;
-  max-width: 160px;
-  padding: 0 20px;
-  background: transparent;
-  transition: all .3s;
-}
-
-#dashboard-header .nav .nav-link:hover,
-#dashboard-header .nav .nav-link.active {
-  background: #fff1;
-}
-
-#dashboard-header .nav .nav-link::after {
-  content: '';
-  display: block;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 3px;
-  background: var(--gradient);
-  opacity: 0;
-  transition: all .3s;
-}
-
-#dashboard-header .nav .nav-link.active::after {
-  opacity: 1;
-}
-
-@media screen and (max-width: 650px) {
-  #dashboard-header .header {
-    height: 15vh;
-  }
-
-  #dashboard-header .header .home-title {
-    background: none;
-    transform: none;
-    margin: 0;
-    padding: 0;
-  }
-
-  #dashboard-header .header .home-title p {
-    font-size: 1.5rem;
-    transform: skew(0deg);
-  }
-}
